refactor(handler): split misc handler files under 300 line limit
