+++ internal/handler/sandbox_templates.go
@@ -36,12 +36,11 @@ type TemplateBuildable interface {
 	DeleteTemplate(ctx context.Context, externalID string) error
 }
 
-// Ensure sandbox.Orchestrator satisfies TemplateBuildable.
 var _ TemplateBuildable = (*sandbox.Orchestrator)(nil)
 
 type SandboxTemplateHandler struct {
 	db       *gorm.DB
-	builder  TemplateBuildable // nil if sandbox orchestrator not configured
+	builder  TemplateBuildable
 	enqueuer enqueue.TaskEnqueuer
 }
 
@@ -53,7 +52,7 @@ type createSandboxTemplateRequest struct {
 	Name           string     `json:"name"`
 	BuildCommands  []string   `json:"build_commands"`
 	Config         model.JSON `json:"config,omitempty"`
-	BaseTemplateID *string    `json:"base_template_id,omitempty"` // UUID of a public template to use as base
+	BaseTemplateID *string    `json:"base_template_id,omitempty"`
 }
 
 type updateSandboxTemplateRequest struct {
@@ -155,7 +154,6 @@ func (h *SandboxTemplateHandler) Create(w http.ResponseWriter, r *http.Request) 
 		tmpl.Config = model.JSON{}
 	}
 
-	// If a base template is specified, validate it's a public ready template and inherit its size.
 	if req.BaseTemplateID != nil && *req.BaseTemplateID != "" {
 		baseID, err := uuid.Parse(*req.BaseTemplateID)
 		if err != nil {
@@ -171,7 +169,6 @@ func (h *SandboxTemplateHandler) Create(w http.ResponseWriter, r *http.Request) 
 		tmpl.Size = baseTmpl.Size
 	}
 
-	// Auto-generate slug from ID after creation
 	tmpl.Slug = fmt.Sprintf("hiveloop-tmpl-%s", uuid.New().String()[:8])
 
 	if err := h.db.Create(&tmpl).Error; err != nil {
@@ -233,324 +230,4 @@ func (h *SandboxTemplateHandler) List(w http.ResponseWriter, r *http.Request) {
 	}
 
 	writeJSON(w, http.StatusOK, result)
-}
-
-// Get handles GET /v1/sandbox-templates/{id}.
-// @Summary Get a sandbox template
-// @Description Returns a single sandbox template by ID.
-// @Tags sandbox-templates
-// @Produce json
-// @Param id path string true "Template ID"
-// @Success 200 {object} sandboxTemplateResponse
-// @Failure 404 {object} errorResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/{id} [get]
-func (h *SandboxTemplateHandler) Get(w http.ResponseWriter, r *http.Request) {
-	org, ok := middleware.OrgFromContext(r.Context())
-	if !ok {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing org context"})
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-	var tmpl model.SandboxTemplate
-	if err := h.db.Where("id = ? AND org_id = ?", id, org.ID).First(&tmpl).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			writeJSON(w, http.StatusNotFound, map[string]string{"error": "sandbox template not found"})
-			return
-		}
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to get sandbox template"})
-		return
-	}
-
-	writeJSON(w, http.StatusOK, toSandboxTemplateResponse(tmpl))
-}
-
-// Update handles PUT /v1/sandbox-templates/{id}.
-// @Summary Update a sandbox template
-// @Description Updates a sandbox template. Resets build status if commands change.
-// @Tags sandbox-templates
-// @Accept json
-// @Produce json
-// @Param id path string true "Template ID"
-// @Param body body updateSandboxTemplateRequest true "Fields to update"
-// @Success 200 {object} sandboxTemplateResponse
-// @Failure 400 {object} errorResponse
-// @Failure 404 {object} errorResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/{id} [put]
-func (h *SandboxTemplateHandler) Update(w http.ResponseWriter, r *http.Request) {
-	org, ok := middleware.OrgFromContext(r.Context())
-	if !ok {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing org context"})
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-	var tmpl model.SandboxTemplate
-	if err := h.db.Where("id = ? AND org_id = ?", id, org.ID).First(&tmpl).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			writeJSON(w, http.StatusNotFound, map[string]string{"error": "sandbox template not found"})
-			return
-		}
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to get sandbox template"})
-		return
-	}
-
-	var req updateSandboxTemplateRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
-		return
-	}
-
-	updates := map[string]any{}
-	if req.Name != nil {
-		updates["name"] = *req.Name
-	}
-	if req.BuildCommands != nil {
-		updates["build_commands"] = commandsToString(req.BuildCommands)
-		// Reset build status when commands change
-		updates["build_status"] = "pending"
-		updates["external_id"] = nil
-		updates["build_error"] = nil
-	}
-	if req.Config != nil {
-		updates["config"] = req.Config
-	}
-
-	commandsChanged := req.BuildCommands != nil
-
-	if len(updates) > 0 {
-		if err := h.db.Model(&tmpl).Updates(updates).Error; err != nil {
-			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to update sandbox template"})
-			return
-		}
-		h.db.Where("id = ?", tmpl.ID).First(&tmpl)
-	}
-
-	// Rebuild if commands changed
-	if commandsChanged && h.builder != nil && tmpl.BuildCommands != "" {
-		go h.builder.BuildTemplate(context.Background(), &tmpl)
-	}
-
-	writeJSON(w, http.StatusOK, toSandboxTemplateResponse(tmpl))
-}
-
-// TriggerBuild handles POST /v1/sandbox-templates/{id}/build.
-// @Summary Trigger a sandbox template build
-// @Description Enqueues an async build job for the template. Poll GET endpoint for status and logs.
-// @Tags sandbox-templates
-// @Accept json
-// @Produce json
-// @Param id path string true "Template ID"
-// @Success 202 {object} sandboxTemplateResponse
-// @Failure 400 {object} errorResponse
-// @Failure 404 {object} errorResponse
-// @Failure 409 {object} errorResponse
-// @Failure 500 {object} errorResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/{id}/build [post]
-func (h *SandboxTemplateHandler) TriggerBuild(w http.ResponseWriter, r *http.Request) {
-	org, ok := middleware.OrgFromContext(r.Context())
-	if !ok {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing org context"})
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-	var tmpl model.SandboxTemplate
-	if err := h.db.Where("id = ? AND org_id = ?", id, org.ID).First(&tmpl).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			writeJSON(w, http.StatusNotFound, map[string]string{"error": "sandbox template not found"})
-			return
-		}
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to get sandbox template"})
-		return
-	}
-
-	if tmpl.BuildStatus == "building" {
-		writeJSON(w, http.StatusConflict, map[string]string{"error": "build already in progress"})
-		return
-	}
-
-	if h.enqueuer == nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "build worker not configured"})
-		return
-	}
-
-	task, err := tasks.NewSandboxTemplateBuildTask(tmpl.ID)
-	if err != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to enqueue build task"})
-		return
-	}
-	if _, err := h.enqueuer.Enqueue(task); err != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to enqueue build task"})
-		return
-	}
-
-	// Update status to building
-	h.db.Model(&tmpl).Update("build_status", "building")
-	tmpl.BuildStatus = "building"
-
-	writeJSON(w, http.StatusAccepted, toSandboxTemplateResponse(tmpl))
-}
-
-// RetryBuild handles POST /v1/sandbox-templates/{id}/retry.
-// @Summary Retry a sandbox template build
-// @Description Deletes the existing snapshot (if any) and starts a new build. Can optionally update build commands.
-// @Tags sandbox-templates
-// @Accept json
-// @Produce json
-// @Param id path string true "Template ID"
-// @Param body body retryBuildRequest false "Optional build commands update"
-// @Success 202 {object} sandboxTemplateResponse
-// @Failure 400 {object} errorResponse
-// @Failure 404 {object} errorResponse
-// @Failure 409 {object} errorResponse
-// @Failure 500 {object} errorResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/{id}/retry [post]
-func (h *SandboxTemplateHandler) RetryBuild(w http.ResponseWriter, r *http.Request) {
-	org, ok := middleware.OrgFromContext(r.Context())
-	if !ok {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing org context"})
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-	var tmpl model.SandboxTemplate
-	if err := h.db.Where("id = ? AND org_id = ?", id, org.ID).First(&tmpl).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			writeJSON(w, http.StatusNotFound, map[string]string{"error": "sandbox template not found"})
-			return
-		}
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to get sandbox template"})
-		return
-	}
-
-	if tmpl.BuildStatus == "building" {
-		writeJSON(w, http.StatusConflict, map[string]string{"error": "build already in progress"})
-		return
-	}
-
-	if h.enqueuer == nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "build worker not configured"})
-		return
-	}
-
-	var req retryBuildRequest
-	if r.ContentLength > 0 {
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
-			return
-		}
-	}
-
-	// Update status to building and reset synchronously to prevent race conditions
-	h.db.Model(&tmpl).Updates(map[string]any{
-		"build_status": "building",
-		"build_error":  nil,
-		"build_logs":   "",
-	})
-	tmpl.BuildStatus = "building"
-	tmpl.BuildError = nil
-	tmpl.BuildLogs = ""
-
-	task, err := tasks.NewSandboxTemplateRetryBuildTask(tmpl.ID, req.BuildCommands)
-	if err != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to enqueue retry task"})
-		return
-	}
-	if _, err := h.enqueuer.Enqueue(task); err != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to enqueue retry task"})
-		return
-	}
-
-	writeJSON(w, http.StatusAccepted, toSandboxTemplateResponse(tmpl))
-}
-
-// @Summary Delete a sandbox template
-// @Description Deletes a template. Fails if agents still reference it.
-// @Tags sandbox-templates
-// @Produce json
-// @Param id path string true "Template ID"
-// @Success 200 {object} map[string]string
-// @Failure 404 {object} errorResponse
-// @Failure 409 {object} errorResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/{id} [delete]
-func (h *SandboxTemplateHandler) Delete(w http.ResponseWriter, r *http.Request) {
-	org, ok := middleware.OrgFromContext(r.Context())
-	if !ok {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing org context"})
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-
-	// Check if any agents reference this template
-	var agentCount int64
-	h.db.Model(&model.Agent{}).Where("sandbox_template_id = ? AND org_id = ?", id, org.ID).Count(&agentCount)
-	if agentCount > 0 {
-		writeJSON(w, http.StatusConflict, map[string]string{"error": "cannot delete template: agents still reference it"})
-		return
-	}
-
-	// Delete the Daytona snapshot if it was built
-	var tmpl model.SandboxTemplate
-	if err := h.db.Where("id = ? AND org_id = ?", id, org.ID).First(&tmpl).Error; err == nil {
-		if tmpl.ExternalID != nil && h.builder != nil {
-			_ = h.builder.DeleteTemplate(r.Context(), *tmpl.ExternalID)
-		}
-	}
-
-	result := h.db.Where("id = ? AND org_id = ?", id, org.ID).Delete(&model.SandboxTemplate{})
-	if result.Error != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to delete sandbox template"})
-		return
-	}
-	if result.RowsAffected == 0 {
-		writeJSON(w, http.StatusNotFound, map[string]string{"error": "sandbox template not found"})
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
-}
-
-type publicTemplateResponse struct {
-	ID   string     `json:"id"`
-	Name string     `json:"name"`
-	Slug string     `json:"slug"`
-	Tags model.JSON `json:"tags"`
-	Size string     `json:"size"`
-}
-
-// ListPublic handles GET /v1/sandbox-templates/public.
-// @Summary List public sandbox templates
-// @Description Returns all public (platform-wide) sandbox templates that are ready.
-// @Tags sandbox-templates
-// @Produce json
-// @Success 200 {object} map[string][]publicTemplateResponse
-// @Security BearerAuth
-// @Router /v1/sandbox-templates/public [get]
-func (h *SandboxTemplateHandler) ListPublic(w http.ResponseWriter, r *http.Request) {
-	var templates []model.SandboxTemplate
-	if err := h.db.Where("org_id IS NULL AND build_status = ?", "ready").
-		Order("name ASC").Find(&templates).Error; err != nil {
-		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to list public templates"})
-		return
-	}
-
-	resp := make([]publicTemplateResponse, len(templates))
-	for index, tmpl := range templates {
-		resp[index] = publicTemplateResponse{
-			ID:   tmpl.ID.String(),
-			Name: tmpl.Name,
-			Slug: tmpl.Slug,
-			Tags: tmpl.Tags,
-			Size: tmpl.Size,
-		}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{"data": resp})
 }
